fix the comparison between int and signed int

+++ src/s2/s2point_span.h
@@ -48,7 +48,7 @@ class S2PointLoopSpan : public S2PointSpan {
   // where each index i >= size() is mapped to i - size().
   reference operator[](int i) const noexcept {
     S2_DCHECK_GE(i, 0);
-    S2_DCHECK_LT(i, 2 * size());
+    S2_DCHECK_LT(i, static_cast<int>(2 * size()));
     int j = i - static_cast<int>(size());
     return S2PointSpan::operator[](j < 0 ? i : j);
   }
